+++ Booleans/a_suspiscious_list.py
@@ -1,0 +1,6 @@
+a = [1, 2, 3]
+b = a
+
+print(id(a))
+print(id(b))
+print (a is b)