+++ Booleans/false_values.py
@@ -1,0 +1,12 @@
+condition = False
+#If you use 0,False, None, or an empty string, list, tuple, or dictionary, it will be evaluated as False.
+if condition:
+    print('This is true')
+else:
+    print('This is false')
+    # False Values:
+    # False
+    # None
+    # Zero of any numeric type
+    # Any empty sequence. For example, '', (), [].
+    # Any empty mapping. For example, {}.