+++ Booleans/log_in_screen.py
@@ -1,0 +1,7 @@
+user ='Admin'
+logged_in = False
+
+if not logged_in:
+    print('Please Log In')
+else:
+    print('Welcome Admin')