+++ Conditionals/Basic_conditionals.py
@@ -1,4 +1,10 @@
-language ='Python'
+language ='Java'
 
 if language == 'Python':
-    print('Conditional was True')
+    print('Language is Python')
+elif language =='Java':
+    print('Language is Java')
+elif language == 'Lua':
+    print('Language is Lua')
+else:
+    print('No match')