+++ Conditionals/log_in_screen.py
@@ -1,0 +1,7 @@
+user ='Admin'
+logged_in = True
+
+if user == 'Admin' or logged_in:
+    print('Admin Page')
+else:
+    print('Bad Credentials')